pyproject.toml
[tool.poetry]
name = "xiaozhi-agent"
version = "0.1.0"
description = "AI agent for ESP32 smart speakers using LangGraph"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
websockets = "^11.0.3"
langgraph = "^0.0.19"
langchain = "^0.0.292"
openai = "^1.1.1"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
asyncio = "^3.4.3"
numpy = "^1.26.1"
python-multipart = "^0.0.6"
httpx = "^0.25.0"
opuslib = "^3.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = ["E203"]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py310"] 